I added a new print_summary test under the test_scripts directory. Additionally, I included Doxygen-style documentation blocks above the safe_fd_printf and account_print_summary functions in account.c, and made a few content updates to both.
While testing handle_login_test, I noticed that test cases 1 and 9 were failing. I modified the case and special characters in the test inputs (e.g., changed hashpass1 to Hashpass1!) to comply with the password policy enforced in account.c. This improved the situation, but occasional failures still occur. I'm not entirely sure, but it might be related to the fact that handle_login() internally calls account_update_password(), which could be affecting some shared or global state. I'll try to figure it out later.